usb: device_next: new USB Video Class implementation
Introduce a new USB Video Class (UVC) implementation from scratch.
The USB Audio Class 2 (UAC2) was taken as inspiration for configuring
the descriptors from the devicetree, for now, only the VideoStreaming
interfaces that describe the supported formats.

It only support a Input Terminal -> Output Terminal chain for now,
but it is possible to add more complex topologies in a compatible
way in the future.

It exposes a native Zephyr Video driver interface, allowing to call
the video_enqueue()/video_dequeue() interface that @loicpoulain did
introduce in 2019 and maintained since then.

The application can query the format currently selected by the host,
but will not be alerted when the host configures a new format, as
there is no video.h API for it yet.

However, there can be video devices declared as phandle in the
devicetree to which the UVC class can then send controls.

Signed-off-by: Josuah Demangeon <me@josuah.net>