drivers: video: gc2145: switch to video CCI library
Reduce the amount of code in the GC2145 driver by switching to the CCI
helpers in video_common.c. The batch I2C write functions are not used
to avoid increasing the size of the ROM, as struct video_reg is 8 bytes
while the ad-hoc struct gc2145_reg is 2 bytes.

Signed-off-by: Josuah Demangeon <me@josuah.net>